+++ cmd/client/client.go
@@ -3,6 +3,7 @@ package main
 import (
 	"flag"
 	"fmt"
+	"github.com/caarlos0/env/v6"
 	"github.com/rs/zerolog/log"
 	"github.com/soerenschneider/dyndns/client"
 	"github.com/soerenschneider/dyndns/client/resolvers"
@@ -26,8 +27,8 @@ var configPathPreferences = []string{
 
 func main() {
 	metrics.Version.WithLabelValues(internal.BuildVersion, internal.CommitHash).SetToCurrentTime()
-	defaultConfigPath := checkDefaultConfigFiles()
-	configPath := flag.String("config", defaultConfigPath, "Path to the config file")
+
+	configPath := flag.String("config", "", "Path to the config file")
 	once := flag.Bool("once", false, "Do not run as a daemon")
 	version := flag.Bool("version", false, "Print version and exit")
 	flag.Parse()
@@ -38,26 +39,32 @@ func main() {
 	}
 
 	util.InitLogging()
-	if nil == configPath {
-		log.Fatal().Msgf("No config path supplied")
+	if *configPath == "" {
+		*configPath = getDefaultConfigFileOrEmpty()
 	}
-
 	conf, err := conf.ReadClientConfig(*configPath)
 	if err != nil {
 		log.Fatal().Msgf("couldn't read config file: %v", err)
 	}
+	if err := env.Parse(conf); err != nil {
+		log.Fatal().Msgf("%+v\n", err)
+	}
+
 	// supply once flag value
 	conf.Once = *once
 	conf.Print()
 	RunClient(conf)
 }
 
-func checkDefaultConfigFiles() string {
+func getDefaultConfigFileOrEmpty() string {
+	homeDir := getUserHomeDirectory()
 	for _, configPath := range configPathPreferences {
-		if strings.HasPrefix(configPath, "~/") {
-			configPath = path.Join(getUserHomeDirectory(), configPath[2:])
-		} else if strings.HasPrefix(configPath, "$HOME/") {
-			configPath = path.Join(getUserHomeDirectory(), configPath[6:])
+		if homeDir != "" {
+			if strings.HasPrefix(configPath, "~/") {
+				configPath = path.Join(homeDir, configPath[2:])
+			} else if strings.HasPrefix(configPath, "$HOME/") {
+				configPath = path.Join(homeDir, configPath[6:])
+			}
 		}
 
 		if _, err := os.Stat(configPath); err == nil {
@@ -65,11 +72,15 @@ func checkDefaultConfigFiles() string {
 		}
 	}
 
-	return configPathPreferences[0]
+	return ""
 }
 
 func getUserHomeDirectory() string {
-	usr, _ := user.Current()
+	usr, err := user.Current()
+	if err != nil || usr == nil {
+		log.Warn().Msg("Could not find user home directory")
+		return ""
+	}
 	dir := usr.HomeDir
 	return dir
 }
@@ -96,7 +107,7 @@ func RunClient(conf *conf.ClientConf) {
 
 	var dispatchers []events.EventDispatch
 	for _, broker := range conf.Brokers {
-		dispatcher, err := mqtt.NewMqttDispatch(broker, conf.Host, fmt.Sprintf("dyndns/%s", conf.Host))
+		dispatcher, err := mqtt.NewMqttDispatch(broker, conf.ClientId, fmt.Sprintf("dyndns/%s", conf.Host))
 		if err != nil {
 			log.Fatal().Msgf("Could not build mqtt dispatcher: %v", err)
 		}